Add docstring to Eccentricity class
