+++ giotto/mapper/tests/test_filter.py
